# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de E-commerce Segura con Swagger Completo",
        "description": "API RESTful para la gestion de clientes, catalogo de productos, ordenes de compra y procesamiento de pagos mediante Stripe. Construida con Laravel 12 y autenticacion basada en tokens (Sanctum).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
...
        "/payments/{order}/status": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Consultar el estado del pago de una orden directamente en Stripe",
                "operationId": "622acc7e763adcdcf4a599c10cf6b7de",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actual del pago"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Orden o pago no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/payments/webhook": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Webhook de Stripe para eventos de pago",
                "operationId": "d4f19bf65ca619a0344ecaaba65c55a8",
                "responses": {
                    "200": {
                        "description": "Evento procesado"
                    },
                    "400": {
                        "description": "Firma del webhook invalida"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar catalogo de productos (acceso publico)",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por pagina (entre 1 y 50)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de productos activos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto (requiere administrador)",
                "operationId": "eddf8c7847239ddc1040daadd814e21b",
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener el detalle de un producto",
                "operationId": "b5099368868dbe7900acd129345eb148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del producto"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto (requiere administrador)",
                "operationId": "2ce632b9d1c0c445557821d09add7ec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto (requiere administrador)",
                "operationId": "76bb7dc4ab3fc3233d5b33ef58e95d7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduzca el token con el formato: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}